# YJIT: Stack temp register allocation