@maximecb 2023-04-04
> In addition to it, they must be spilled for jit.peek_at_stack

As a potential improvement in a future PR, it might be possible in that case to write the values to the stack but preserve the register mapping. As in, the values need to be written to the stack for peek_at_stack, but they don't need to be unmapped from registers?